add button to open data dir
