+++ harbor-ui/Cargo.toml
@@ -24,6 +24,6 @@ bitcoin = { version = "0.32.4", features = ["base64"] }
 bip39 = "2.0.0"
 fedimint-core = "0.5.0"
 fedimint-ln-common = "0.5.0"
-opener = "0.7.2"
+opener = { version = "0.7.2", features = ["reveal"] }
 serde = { version = "1.0.217", features = ["derive"] }
 serde_json = "1.0.138"
