+++ harbor-ui/src/components/mod.rs
@@ -65,3 +65,6 @@ pub use absolute_overlay::*;
 
 mod confirm_modal;
 pub use confirm_modal::*;
+
+mod debug_stuff;
+pub use debug_stuff::*;
