+++ harbor-ui/src/routes/settings.rs
@@ -2,8 +2,9 @@ use bitcoin::Network;
 use iced::widget::{column, pick_list, text, Checkbox};
 use iced::{Element, Padding};
 
-use crate::components::{basic_layout, h_button, h_header, menu_style, pick_list_style, SvgIcon};
-use crate::components::{Toast, ToastStatus};
+use crate::components::{
+    basic_layout, debug_stuff, h_button, h_header, menu_style, pick_list_style, SvgIcon,
+};
 use crate::{HarborWallet, Message};
 
 pub fn settings(harbor: &HarborWallet) -> Element<Message> {
@@ -41,39 +42,8 @@ pub fn settings(harbor: &HarborWallet) -> Element<Message> {
     })
     .menu_style(menu_style);
 
-    let add_good_toast_button =
-        h_button("Nice!", SvgIcon::Plus, false).on_press(Message::AddToast(Toast {
-            title: "Hello".to_string(),
-            body: Some("This is a toast".to_string()),
-            status: ToastStatus::Good,
-        }));
-
-    let add_error_toast_button =
-        h_button("Error Toast", SvgIcon::Plus, false).on_press(Message::AddToast(Toast {
-            title: "Error".to_string(),
-            body: Some("This is a toast".to_string()),
-            status: ToastStatus::Bad,
-        }));
-
-    let test_confirm_modal_button = h_button("Test Confirm Modal", SvgIcon::Shield, false)
-        .on_press(Message::SetConfirmModal(Some(
-            crate::components::ConfirmModalState {
-                title: "Test Modal".to_string(),
-                description:
-                    "This is a test of the confirm modal. Are you sure you want to proceed?"
-                        .to_string(),
-                confirm_action: Box::new(Message::Batch(vec![
-                    Message::AddToast(Toast {
-                        title: "You confirmed!".to_string(),
-                        body: None,
-                        status: ToastStatus::Good,
-                    }),
-                    Message::SetConfirmModal(None),
-                ])),
-                cancel_action: Box::new(Message::SetConfirmModal(None)),
-                confirm_button_text: "Confirm".to_string(),
-            },
-        )));
+    let open_data_dir_button = h_button("Open Data Directory", SvgIcon::FolderLock, false)
+        .on_press(Message::OpenDataDirectory);
 
     let column = match (harbor.settings_show_seed_words, &harbor.seed_words) {
         (true, Some(s)) => {
@@ -91,15 +61,20 @@ pub fn settings(harbor: &HarborWallet) -> Element<Message> {
             let button = h_button("Show Seed Words", SvgIcon::Eye, false)
                 .on_press(Message::ShowSeedWords(true));
 
+            let debug_stuff = if cfg!(debug_assertions) {
+                Some(debug_stuff())
+            } else {
+                None
+            };
+
             column![
                 header,
                 button,
                 onchain_receive_checkbox,
                 network_list,
-                add_good_toast_button,
-                add_error_toast_button,
-                test_confirm_modal_button
+                open_data_dir_button,
             ]
+            .push_maybe(debug_stuff)
         }
     };
